Not really a gap limit, rather a look-ahead
The gap limit is the number of consecutive "unused addresses"
after which it's safe to assume that no more addresses will be
used in any transaction. This type of sequential scanning is only
feasible with a local full UTXO set. By knowing in advance when
to stop generating the deterministic addresses it's easy to build
a Bloom filter that won't change until sync is complete.

However, we don't store the UTXO set and we don't want to rely on
a 3rd party database or web service to scan used addresses in
advance. Our approach is:

- Download blocks.
- Register wallet transactions from blocks.
- If a transaction ended up consuming all the look-ahead
  addresses: generate new addresses, rebuild Bloom filter, request
  upcoming blocks again with the new filter.

There's no gap limit concept in this process. Still, the look-ahead
is an essential parameter to find all relevant transactions.