Fixed non-extending fork
The returned nil addedBlock in case the block is extending a fork
rather than main chain, caused the following methods in
WSBlockChainDownloader:

- appendBlockHeaders:error:
- appendBlock:error: (TODO)
- appendFilteredBlock:withTransactions:error:

to halt with a "Header not added" or "Filtered block not added"
debug messages. Instead, a block was actually added. Especially
in appendBlockHeaders, the erroneous code caused the loop to stop
and discard further headers, thus preventing the blockchain from
reorganizing.